[virtualization] Gates calls to connect to vsock.
server_daemon_test does not wait for a vsock connection to
be made before attempting to use it.  If the socket is used
before it is ready, writing to it fails resulting in a test
flake.  This change waits for the call to connect to the
vsock to complete before continuing with the test.

In testing this patch, it was discovered that the
guest_interaction_daemon being alive does not guarantee
that the program is listening on the vsock.  If the daemon
is not listening on the vsock, the call to connect will
fail.  This patch adds a log after the call to listen.
Tests can then query journalctl over the guest's serial
console to determine the state of the daemon before
attempting to connect.

Bug: 36321
Test: fx run-test guest_interaction_tests -t \
      server_daemon_test
      fx run-test netemul_guest_test
Change-Id: I0944baa16f6a77f6a82166fd704cbd9c1eea8954
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/365433
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>