[storage] Update blobfs metrics.
These metrics are only printed out at filesystem unmount so this should
not affect any server metrics.

Removes counting the pre-cache of blobs as a "lookup." Originally the
lookup metric tracked blobs as they were opened. Later a pre-caching
layer was added in Blobfs::InitializeVnodes() and this metric was added
there.

Combining both a count of all blobs in the system plus all blobs opened
doesn't make sense. This removes the lookup metric counting from the
pre-caching so that the lookup metric counts only blobs accessed via
Directory::Lookup. If we need more filesystem-wide blobfs metrics, those
should be added in a separate value.

Removes the "writeback thread" metrics. These were never updated and
reflected the design before there was a journal. If we want this type of
data it would be a journal statistic we would need to add in separately.

Timings are only taken when Cobalt metrics are enabled. Since otherwise
the only use of these statistics is a printout to the console that users
will never read, this behavior is preserved. To avoid being confusing,
printing timings are disabled to avoid confusion.

Bug: 52351

Change-Id: I73228d7cc067b04b5ee8ef07bf1280545d8783c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391496
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>