[sl4f][process_trace] Fix build error for process_trace
The build rule makes an implicit library for the binary and was using
the same directory as the root for both the sl4f client library and the
implicit library, causing conflicts. This change results in the root
changing from client/ to client/bin/, avoiding the conflict.

TEST: Include //sdk/testing/sl4f/client:process_trace and fx build
BUG: 52435
Change-Id: I5c20e7328f61eebd906f8324500454a4956a9ac1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391499
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>