[audio_core] Flush queue on unlink
Fix a use-after-free error condition that can occur if there is a
mix-job in progress when a renderer is removed from the system. When
this happens, the outputs mix thread may be holding a reference to
in-flight packets while RouteGraph is performing unlinks and removal
of the Renderer. In some situations, the Renderer can be free'd while
the mix job is still retaining packets (and a packet queue) from that
Renderer.

This is problematic because we use a SlabAllocator, which is owned by
the Renderer, to allocate those Packets. So when this happens we leave
pointers to invalid memory in the mixer thread.

We resolve this issue by running a Flush on the link's PacketQueue
during unlinking and passing a strong reference to the Renderer to the
Flush callback.

Fixed: 49622
Test: fx test audio_core_unittests
      New test fails with an ASAN violation without the patch to
      base_renderer.cc
Change-Id: I8fe434aff5d0bf558ba27e12263edb091346e8e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391595
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>