[i18n] Adds a time correctness system test
This test will fail if the Dart SDK configuration changes such that the
dart timezone is broken on a running device.  It should prevent the "OMG
the clock is showing the wrong time" which sometimes happened -- or, at
least, it will prevent such cases where there is a discrepancy between
the system timezone and the runner timezone.

This test has been a long time in the making.  In order for this test to
be able to pass on an actual device, the following changes were needed
(some by me some by others):

- Adding a global component context to dart VM.
- Adding dart VM inspect
- Fixing the basemgr wire-up of locale context bug, which caused the
  locale change to be invisible in the dart VM.
- Fixing the Fuchsia flutter roller.
- Fixing a certain misplaced `assert` statement.

Bug: 43341
Change-Id: I3ebc3d74b408f828c32b0bb282f88a46e15f66eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/367445
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>