[media][audio_core] Enable input driver_gain_db
On driver initialization, driver_gain_db is applied to the input driver.
The driver_gain_db value is sourced from the device config; if the key
is missing from the config or the config is not available, a default
gain value of 0.0db is applied.

Bug: 51003

Test: fx test audio_device_tests
Change-Id: Ie65781ccb130e5587a6bb42fdf419286a7a06458
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391422
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Payton Bradsky <pbradsky@google.com>