[inspect][rust] Add weak node references.
Weak references is relevant primarily in cases where nodes or
properties are added post-instantiation. For instance, consider a
type that has an `Option<StringProperty>` that is attached to the
parent. If initialized to `None` upon construction (when parent
is usually provided), there is no way to attach to inspect at
a later point, without providing parent in the function chain on
non-constructor regular method calls.

For `fuchsia_derive_inspect`, this unlocks having `Option<T>` be
represented as a node or property that is absent when `None`.

This change moves the custom drop-logic to the innermost type,
the `InnerRef`. Since deallocation depends on the node or property
type, a new trait `Free` with different implementations are provided.

Fixed: 51720
Change-Id: I9d8bd5674a535b94559b6bd6be718d9ca9ab7eb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390853
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Didrik Nordstrom <dnordstrom@google.com>