[inspect] Refactor NodeHierarchy serialization
This first pass implements Serialize on NodeHierarchy directly.
The reason it was done separately is purely historical when
NodeHierarchy used to be in a different crate.

This also makes the `Schema` in Archvist generic and not
depending on the format. This simplifies introducing new formats such
as CBOR, FIDL Text, IQuery Text, etc.

Additionally, the Schema is now
generic over the Key. This makes it easily extensible to support Logs
that have a different Key type than Inspect and Lifecycle.

Unfortunately, deserialization is a bit more complicated given that
it's over-fitted to json at the moment and simply manually implementing
Deserialize is by far not as straightforward as implementing Serialize.
A follow up will refactor it so that we have
`impl Deserialize for NodeHierarchy`.

Tested: just a refactor, existing tests continue to pass

Change-Id: I6e4238ac55d4e376cee64ff81dd7f8148b2ce750
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391560
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Luke Nicholson <lukenicholson@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>