[region-alloc] Relax RegionPool rules.
Previously, to use a RegionAllocator, you were required to use a
RegionAllocator::RegionPool object as the SlabAllocator based source
of bookkeeping objects.  After this change, these rules are relaxed.
A RegionAllocator may still be used with a Pool, but it can also
operate without one, simply allocating from the heap.

This is preparation for a new use in the kernel where the allocator
will be used mostly for testing for attempts to create resources and
MMIO mappings which could conflict with a set of forbidden ranges.
The RegionAllocator provides a convenient (and well tested) API for
doing stuff like this, but using a RegionPool is way overkill in this
case.  The forbidden regions will be set up once at startup, and then
never changed.  Minimizing heap fragmentation because of bookkeeping
churn is not a concern in this case, and potential overallocation
because of slab allocation is undesirable.

Tests: new unit tests.
Bug: 34029

Change-Id: I0ae174d02621ce644be113c4b55297b194b6c1ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390795
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Christopher Anderson <cja@google.com>