[modular] Fix flake in new agent_runner_unittest
When the AgentController channel is closed, it is similar to sending a
final message on the channel. Before a Binding or InterfacePtr will
return !is_bound(), the run loop whose dispatcher is waiting on that channel must have a
chance to consume and process the close message.

Hypothesis: There were two tests that were not waiting for this to happen, checking
if is_bound() was false based on a synchronization point that could
arrive before the close messages did.

I could not reproduce the original flake with n=10000 retries. This
change still works and the tests still do not flake with n=10000
retries. I suspect this will fix the underlying race condition.

Fixed: 49609
Change-Id: I828f06149b66423027f0ef5c29e116b3105d8829
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392237
Commit-Queue: Gabe Schine <thatguy@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Testability-Review: Gabe Schine <thatguy@google.com>