[shell] implement shutdown
This allows the client to destroy the interpreter it created. When the
interpreter is destroyed, it checks that everything is deallocated
correctly (there is no memory leak).

Fix a memory leak for String when the string was created using a
StringContainer.

Change-Id: I2785051344ec65daabdc89cd72fdb563abbcd2eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390585
Commit-Queue: Vincent Belliard <vbelliard@google.com>
API-Review: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>