[kernel][x86] idle_thread: Spin a little bit before HLT
The fuchsia idle thread is scheduled whenever there is no immediate
work to execute on a CPU.

On CPUs without MONITOR/MWAIT (most importantly VMs), spin a little
bit before executing the halt instruction. The latency of a halt
may be pretty long, so a short spin can avoid that on receivers when
there are many core sleep/wakeup events within a small window.

Halts may be particularly expensive in nested VMs, as both an L0 and
L1 hypervisor are in the path, so avoiding them for short sleeps may
help.

This CR/ is a no-op on CPUs with MONITOR/MWAIT support.

This is a short-term fix to reduce test runtime; longer term, there
are two follow-ups that would be useful:
1. This code could elide the IPI on the sender side if a receiver
   is spinning. That would save a fair amount of work on a waker
   and on the receiver, particularly in a VM.

2. Generalize the C-state tables in the monitor/mwait branches
   to support non-MWAIT sleep modes ('spin' and 'halt'), so that
   the sleep predictor can pick from "spin", any available mwait
   state, or halt as appropriate for the next deadline and the
   predicted sleep time.

   We want to do this for a number of other reasons (better AMD
   and Intel Apollo Lake support; cases where C1 latency is longer
   than the predicted sleep time/deadline); and it should clean up
   mp.cc a little bit.

Tested:
1) Runtime of
   fuchsia.com/fidl_compatibility_test#meta/fidl_compatibility_test.cmx
   in a nested VM on GCP dropped by ~8%.

   Before (seconds):
   57.6, 57.6, 57.2, 59.1, 60.0, 58.6, 57.1, 58.8, 59.1, 59.9

   After (seconds):
   53.2, 53.5, 54.0, 53.1, 52.6, 52.8, 52.8, 53.7, 55.0, 55.5

Bug: 49133 profile fyi builders timing out with ...
Bug: 33981 Fuchsia should run on Intel NUC6CAYH
Bug: 37450 Fuchsia should run well on AMD Ryzen (17h), ZP-B1 and ZP/PiR-B2
Bug: 33638 Fuchsia should run (well) on grunt (AMD) chromebook
Change-Id: Ieca42e7482d6906f3a7a2593fad7523a12876f8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383633
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>