[region-alloc] Add intersection/containment test operation
Add the ability to test a user defined region against the set of
either allocated regions, or available regions, to see if the user's
region either intersects any of the regions in the chosen set, or is
completely contained by any region in the chosen set.

This makes it easy to use the region allocator as type of region
allow/deny list.  We can deny a region if it intersects any of the
regions in a forbidden set, or we can allow a region if it is
completely contained by any of the regions in the allowed set.

Tests: new unit tests
Bug: 34029
Change-Id: I98c631cb512eaf22b37814b93b6b3b509e9ead4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390796
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Christopher Anderson <cja@google.com>