[intl] Fixes a flake in intl_services_test
While I'm not certain of the exact mechanism, the flake was caused by
the binding set being empty on the server side at the time the
`OnChange` event was supposed to be delivered.  This meant that the
test never received its `OnChange` notification so never progressed
beyond the point where it waits for it to arrive.

I suspect the reason for this was that we built a callback setup first,
then called `RunLoopUntil` which processes the binding to the server
among other things.  And if the binding happened to be processed after
the OnChange was supposd to be sent, we miss an OnChange event.

Resolved by introducing a dummy call to GetProfile at the very beginning
of the tests, which ensures that the server is aware of a client's
existence, before we start the "real" tests.

The problem was rare in regular execution, but by attaching `fx fidlcat`
to the test as shown below it became almost 100% reproducible. I guess
the watched programs became slow enough that the data race became
prominent.

```
fx fidlcat --remote=name="intl*"
```

Multiply: intl_services_tests: 80
Fixed: 50789
Change-Id: I62b35360d9a10effaccdabeb70d541d6efb3d9c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392776
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>