[fidl-lint] Enable color output in fidl-lint
Currently, fidl-lint does not explicitly enable color output in any
case. However, it was sometimes erroneously enabling it due to a bug.
fidl::reporter::Format takes a bool `enable_color` and also a parameter
`squiggle_size` with a default value. fidl::utils::FormatFindings was
passing in `squiggle_size` for `enable_color`, which silently compiled
because it is a size_t reinterpreted as a bool, and because the next
param has a default value.

This CL fixes that bug and also enables color output in fidl-lint
depending on the presence of the NO_COLOR env var and whether isatty
is true.

Change-Id: Id048673278500bb8766d8e9fbff3151f1cb7dfac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392793
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>