[camera] Fix threading race in camera-gym BufferCollage.
This change fixes a race in the BufferCollage.AddCollection method.
Previously, the method made fidl API calls on the caller's thread, even
though the bindings were associated with the class's thread. This is
not supported by the hlcpp fidl bindings and can lead to crashes. This
change puts the bulk of AddCollection into a single task, and removes
the use of all but one fit::promise, opting for the more conventional
nested callback approach. This has a side-effect of serializing the
calls to Scenic and Sysmem, but this should not impact performance.

Test: Ran camera-gym and verified no behavior change.
      Ran camera-gym with cherry-picked fxr/385453.
Change-Id: I46c2717937c8e089921d0ca803af9b303367df30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392261
Reviewed-by: Ernest Hua <ernesthua@google.com>
Testability-Review: Matt Sandy <msandy@google.com>
Commit-Queue: Matt Sandy <msandy@google.com>