[kernel][vm] Remove racy asserts
These asserts that confirm that a vm_page is in a list are well
intentioned, but ultimately racy. When these asserts are performed the
queue_node is owned by the page queue, and any usage of it needs to be
serialized with the page queue lock.

Although the page is guaranteed to be in a list, the page queue could
be moving the page between lists, whilst holding its lock. This makes
performing this assert, without holding the page queue lock, ultimately
racy.

Test: Exploiting this is difficult to do reliably as it requires
precise timing between multiple threads, so no tests added.

Bug: 52716
Change-Id: I304eda4580d81688f88bf3acf5dbb7adedb11425
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392496
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>