[paver] Add mechanism to use bufferred client
This CL impelments a mechanism to swtich to the new SyncClientBuffered.
The change consists of the following:

1. A paver context object is introduced in the paver service to host
and share information and unique instance of objects among created
partition-clients.

2. A unique instance of SyncClientBufferred is created and hosted in the
paver context object. The context is passed to a new partition client--
AstroSysconfigPartitionClientBuffered for performing IO.

3. AstroPartitioner queries boot argument astro.sysconfig.io-optimization
to determine whether or not to use
AstroSysconfigPartitionClientBuffered.

Bug: 47505
Change-Id: Id6b39b559cd0b2099ba94f3069c2f2de669148cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382502
Commit-Queue: Yecheng Zhao <zyecheng@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: David Pursell <dpursell@google.com>