[kernel][ktl] Prepare __external_threading header for upstream changes
This is the first part of a three-stage soft transition.  The
upstream libc++ headers have changed so that their requirements
of the supplied <__external_threading> header are incompatible
with the previous ones.

This updates the header for the new API but leaves the old state of
affairs conditionally defined for the transition.  This will permit
rolling out the new toolchain with the new libc++ headers, which is
the second stage of the soft transition.  This is the last known
blocker to landing a new toolchain, so this transition is expected
to complete within a few days at most.

The third and final stage will remove the temporary code under `#if`
so that ktl is compatible only with the new upstream libc++ headers.

It so happens that since the last toolchain update landed in
Fuchsia, the upstream <version> header has added a new
feature-test macro `__cpp_lib_span`.  Though this is wholly
unrelated to the header changes requiring this update, it can
serve as a convenient proxy to distinguish the old set of headers
from the new set.

Bug: 52189
Change-Id: I58b7de663df3df94cf299a212e988fa4982ae93b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392209
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>