[debugger] Clarify symbol status.
Updates the symbol index when sym-stat is called. This ensures the index
stats are up-to-date with any modifications to the symbol path.
Previously, updates to the symbol-path was not reflected in the status
until something was loaded, which made diagnosing problems more
difficult.

Add text to the output of sym-stat indicating that the index was just
updated so there's no ambiguity.

Add error output to the symbol load status in all paths. Previously
there were some paths in loading files in the build ID index that didn't
write status.

Add more troubleshooting to the help.

Change-Id: I37003520998f090be0a25d070057848ac4e41f51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375918
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Brett Wilson <brettw@google.com>