[debugger] Don't escape C++ destructor names.
Destructors were erroniously escaped for display, as in
"MyClass::$(~MyClass)" This is because the formatting code that checks
for escaping was different than the tokenizing code that knows about how
to parse names.

This change adds a new helper function to the tokenizer for the name
formatter to use to know whether something is a valid name and therefore
doesn't need escaping.

"$" was removed as a valid name character. Previously this was in there
but the case never got hit because "$" is handled as a special-case
before this code was called. Without this change.

Change-Id: Icc31cbf2921bc30252c98962fd7a640a2ebc07c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392416
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>