[SL4F] allow graceful exit of a component
The code was treating any component exit as a failure. This change accepts a
result code of 0 (widely accepted as a graceful shutdown/exit) as a non-error.
A non-zero exit code will continue to be reported as an unexpected exit.

Bug: 52840
Test: fx run-e2e-tests sl4f_test
Change-Id: I08f16821d53d352fdd4a0eebdeabf136a7b3e8d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391754
Reviewed-by: Tom Turney <tturney@google.com>
Reviewed-by: Judy Cheng <chengjudy@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Testability-Review: Tom Turney <tturney@google.com>