[cml] Introduce static EventStreams
We will be using the CapabilityRouted v2 event in order to do log
attribution of v2 components. However, a component can connect to
a LogSink before Archivist starts and so we must ensure that Archivist
starts in response to a CapabilityRouted event.

The way to do this is to enable starting a component in response to an
Event. Static EventStreams enable this.

This CL also deletes the golden tests which are exact
whitespace-for-whitespace tests which doesn't mesh well with formatting
of cm files.

Sample syntax:

{
    event_stream: [
        "began",
        "capability_ready",
        "destroyed",
    ],
    as: "/svc/my_stream",
},

Bug: 51179
Change-Id: Iae0b089b05c85e2f4609de13ae0530c3eaa39bcc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390614
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>
API-Review: Gary Bressler <geb@google.com>