[zircon][tests] Do not assert on thread time not changing
The signal that a thread is terminated is returned before the thread's
final reschedule and accounting. It is not correct to assert that a dead
thread's cpu time does not change, since we do not make strong
guarantees about this ordering.

Fixed: 52870
Bug: 52872

Change-Id: Icd9f3453217c3c392dd0c12d4948ae0f0e4f54b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393090
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>