[power_manager] Add DriverManagerHandler node
Implement a new node called DriverManagerHandler that is responsible for
serving the fuchsia.power.manager.DriverManagerRegistration service. The
Driver Manager will be the sole client of this protocol and will use it
to provide the Power Manager with the protocol instances that it
requires in order to drive orderly system shutdown. The new node will
not be instantiated in the Power Manager until it is specified in a node
config file.

Bug: 51043
Bug: 51078
Bug: 51092
Test: fx test power_manager_tests
Multiply: power_manager_tests
Change-Id: I6f7d8e03346c5e65adfd77e79d4a6a68fb09821f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391681
Testability-Review: Patrick Shickel <pshickel@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>