[blobstats] Made blobstats a proper host tool.
Initial refactoring of the blobstats code, starting with
building it as a host tool. That involves passing all Dart
static analysis, which was ignored before when blobstats
was run as a standalone script.
Also added the first couple of tests, now that the code
allows to add tests. At the same time, removed the empty
blobstats_host_test which is now replaced by the end-to-end
fx test added in the last CL.
Tested by running `fx blobstats` in addition to the unit tests
and the end-to-end test.

Change-Id: I851e778c637c93dbe6f4a8621d22522cfba7c9d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391073
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Claudio Cherubino <ccherubino@google.com>