[brcmfmac] Add MSGBUF ioctl support
This commit adds support for MSGBUF commands in the ioctl format.  This
is a command format derived from the Linux history of the driver; the
name "ioctl" is itself fairly meaningless on Fuchsia.

Functionally, this provides the implementations of set_dcmd() and
get_dcmd() in the brcmf_proto interface.

This commit also fixes a race in DmaPool::Buffer between buffer
releasing and re-acquiring, in the common use case that a thread passes
a buffer off to another thread before releasing it.

Bug: 40719
Test: Booted and fx run-test brcmfmac_msgbuf_test on NUC
Change-Id: I120244cdad706fc8c39524da4eed372991570f6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/340889
Commit-Queue: John Sheu <sheu@google.com>
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>
Testability-Review: Louis Yung-chieh Lo <yjlou@google.com>