[archivist] Fix dropped logs due to contended lock
If the component stats log is contended, the poll state of the lock was
returned as the poll state of the log stream. This would then cause the
log stream to be terminated early, and drop logs.

This state can occur when there is a large number of requests at once,
and all requests share the same source identity. This was true of all
driver log sink requests, that occur before archivist is started.

This change moves the stats handling outside of
LogMessageSocket::poll_next.

Bug: 49457
Bug: 52619
Change-Id: I827c9dc35c2ea371055fc656b342ad11c44243f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392973
Testability-Review: Abdulla Kamar <abdulla@google.com>
Testability-Review: Adam Perry <adamperry@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>