[network-manager] Return futures to the eventloop
Let the eventloop explicitly join on the futures returned by workers
instead of spawning the workers' futures straight to the Executor. This
allows the eventloop to observe workers' errors (and potentially take
action in the future).

This change nows result in NetworkManger exiting with an error if any of
the workers terminate early, with the exception of the Overnet worker as
the Overnet service may not be available in all builds. Other than this,
there are no behaviour changes.

Test: Existing tests pass.

Change-Id: I9e7c8bbb096c0e25615fbcab28f46cf97538a285
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392330
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>