[hwstress] Avoid status lines being overwritten on crash.
Currently, we write status lines as follows:

  "foo\r"

where "\r" will cause the cursor to move to the left-hand-side, allowing
us to replace it later.

Unfortunately, this means on a binary crash, the most recent status line
is lost, because the shell prompt (or any other output to stdout) will
overwrite it.

Update the status line to emit the "\r" when clearing a line instead,
meaning that the most recent status will still be shown on a binary
crash.

Bug: 51547
Change-Id: I9134ef1b2504d64f39f419fecd621131b571c667
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392975
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>