[hwstress] Allow CPU workload functions to determine when to exit.
Currently, the |CpuStressor| class calls a workload function in a tight
loop. This makes it hard for the workload function to perform any setup
or teardown.

We adjust the workload function API to accept a "StopIndicator", and
make the workload function itself responsible for determining when to
stop.

This allows for workload functions to perform set up / clean up, such as:

  void DoWork(const StopIndicator& indicator) {
    // Set up
    void* foo = malloc(...);

    // Main loop.
    do {
      // (do work)
    } while (!indicator.ShouldStop());

    // Clean up
    free(foo);
  }

Bug: 51547
Change-Id: Ia4e7be3e695ef7de836515f47c5c44eabe1c024c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392978
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>