[fidlc] Handle diagnostic spans indicating EOF in JSON output
Currently, diagnostics with a span indicating the end of source file
cause an error when fidlc is invoked with --format=json. The JSON
diagnostic printer includes both the start position and end position of
a span, unlike the text printer. On a span that points at the end of
the file and that is of size 1 (it points to the null-terminating
character), this causes fidlc to read past the end of the string to get
the end position of the span.

This CL makes an exception in diagnostic printing for spans that begin
at the end of the source file.

Test: fidlc host tests
Change-Id: I83cf991836462f4f4187b3d930230912ce129075
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393271
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>